+++ azure/container-linux/kubernetes/workers/workers.tf
@@ -1,6 +1,7 @@
 locals {
-  # Channel for a Container Linux derivative
   # coreos-stable -> Container Linux Stable
+  # flatcar-stable -> Flatcar Linux Stable
+  flavor  = split("-", var.os_image)[0]
   channel = split("-", var.os_image)[1]
 }
 
@@ -24,10 +25,21 @@ resource "azurerm_linux_virtual_machine_scale_set" "workers" {
   }
 
   source_image_reference {
-    publisher = "CoreOS"
-    offer     = "CoreOS"
+    publisher = local.flavor == "flatcar" ? "Kinvolk" : "CoreOS"
+    offer     = local.flavor == "flatcar" ? "flatcar-container-linux" : "CoreOS"
     sku       = local.channel
     version   = "latest"
+  }
+
+  # Gross hack just for Flatcar Linux
+  dynamic "plan" {
+    for_each = local.flavor == "flatcar" ? [1] : []
+
+    content {
+      name = local.channel
+      publisher = "kinvolk"
+      product = "flatcar-container-linux"
+    }
   }
 
   # Azure requires setting admin_ssh_key, though Ignition custom_data handles it too